🐛 make sure providers are not killed after an asset is scanned (#1663)
This makes sure that scans for a list of assets work. The current
implementation would kill the provider after the first asset is scanned.
We do a simple fix here that will only kill open providers after the
whole scan is complete. As a long term fix we should probably look into
keeping track of assets that need a specific provider. Only when all
assets for the provider have been scanned, it can be killed.

Signed-off-by: Ivan Milchev <ivan@mondoo.com>