🐛 fix detection of provider via connector
We have conflated connection types (which are coming from asset detection) and connection names (which are coming from CLI). One of the last changes has broken the support of CLI-based connector-detection (at least when the connection type and name differed, like in the case of OS).

This PR clarifies the difference between these types, by allowing the `EnsureProvider` call to either take the type or the name for the connection. In the CLI, we get names like `docker`, which then translate to the `os` provider. In contrast, the connection type is something like `docker-image` which translates to the `os` provider. We could have mixed them all together, but it may lead to confusing results (it should not, but it can and then it really falls apart).

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>