🧹 testutils: load schema from LR (#1670)
We don't have all the resources json schemas in this repo. That's
because they get generated together with the build process of each
provider. Because we don't all the schemas in the repo, testutils cannot
just be included in other projects, because they don't go through the
json resources schema generation.

We now have 2 options: (1) check in all the schemas into the repo or (2)
generate the schema from LR files.

This PR takes the second option. It treats these LR files in the same
way the schema loader treats JSON files: they contain all the metadata,
we don't need to generate any Go-files for tests, we don't need to
generate manifests, we only need the schema which is fully
self-contained in LR files. Thus, instead of parsing JSONs step by step,
it parses LR files step by step (but natively) and get the schema from
them. This reduces the amount of files that will make it into regular PR
reviews by default (because the auto-generated schema doesn't require
reviews).

This may change in the future and we are very open to discussions on
this decisions.

Required for cnspec v9.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>