🧹 remove mqlc.MustCompile
This method also exists in explorer/executor, but with a few differences:
1. it has tests
2. it is actually in use (in any of our projects)
3. it provides useful functionality like using the default runtime

This method is used as an analogy to regexp.MustCompile and thus works especially well in tests and default code paths.

However, this needs a bit more follow-up work, because the runtime may not contain the providers necessary for the piece of code to run. This is surfacing right now in cnspec, where MQL code is requested by the reporter which is provided by the `os` plugin (which may not be installed!). Thus, these methods will likely need a bit more work...

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>