🐛 fix mondoo.capabilities
1. the detection was broken and crashed and
2. it pulled content from the os provider (sacrilege! luckily only a small file but still)

There were a few options to approach this, but trying to detect this dynamically or shoving it into untyped metadata was not a good approach. Instead we now have a field attached to the connection config which can transmit capabilities per connection.

We still only expose this via the `mondoo.capabilities` field which should be deprecated and replaced with a cleaner approach.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>