🐛 os.Connect should not return the asset in the inventory
The behavior of the connect call is defined in the proto here: https://github.com/mondoohq/cnquery/blob/main/providers-sdk/v1/plugin/plugin.proto\#L35-L38
It clarifies that the inventory only returns other discovered assets, while the asset field contains the one asset that the connection call was requesting.

Remove the code that added the connection asset to the inventory, since it does not count as a discovered asset.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>