🐛 support bridge-resources (#1678)
These are resources without providers, e.g. `my.name.is.bob` as a
resource creates the bridge resources `my`, `my.name` and `my.name.is`.
These cannot be created in the provider that created `my.name.is.bob`
because they are only bridging in nature. That said, users should still
be able to type `my.name` and not be greeted by crimson letters
desperately proclaiming an error. Not helpful.

This PR not only adds support for this mechanism but also adds more
testing to LR generation.

(I ran into this with the `platform` resource, which is on the way to be
replaced long-term anyway...)

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>