🐛 fix aliased resources
We could no longer call aliased resource properly. This was partly because the way the schema generation had changed (where I missed to migrate the `IsImplicitResource` field) and partly because the resource lookup had also changed (since we want to call the base resource now with the provider, not the alias).

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>