🔧 add configuration for providers in dev env
This has now happened a few too many times: You get a PR with a change to a provider, but forgot to update your builtin.go file. After 30min you realize you have been testing the providers on disk, instead of the ones in the repo. Also configuring providers is a bit annoying (with all the things you need to change in `builtin.go`).

This is now streamlined in `make providers/config`. You can configure it in `providers.yaml` in the root folder.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>