🐛 support Never in json read/write
Serializing to string tries to follow timestamp norms. However, the two infinity times are set as beyond regular date ranges and cause it to crash.

We decided to use integers to represent the unix time (which is intuitive from reading it) and is much faster to read/write.

However, due to limitations in JSON precision for integers (RFC 7159) we had to limit the range and eagerly interpret things to be either Never-times (future and past).

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>