🧹 use matrix for provider build (#1806)
superseeds #1795

This moves the provider build to a matrix build and also adjust the
build script to use static compile. There is followup work required to
align the bash script with the make script a bit more.

In addition it removes the latest json configuration from the build
process since it was error prone. It will be generated in releasr to
make the generation more stable.