🐛 support sudo over ssh on alpine
stat does not support all the flags, especially printf. We can skip newline escapes, since we only grab simple fields. The only limitation here is the SElinux context, which we do not return yet (so we can revisit that then). Even so, it is the last value returned, so we can stick with `-c`

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>