🐛 fix error when cloning resources
Mandatory arguments were never submitted, because the new provider concept meant that we had to request all the fields individually. We were missing a good pattern to execute this.

Turns out the CreateResourceWithID was only used in one place and its primary purpose was to clone the resource. This method has now been adjusted to better reflect its purpose and its arguments can now do the heavy lifting or getting all the necessary fields for cloning the resource. This method can be further refined to grab mandatory fields automatically.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>