🐛 avoid crashing terraform on unsupported plan calls (#1944)
Happens when we call plan resources without operating on a Terraform
plan. This needs better handling for unsupported or invalid data. I
strongly recommend we instead inform users that there are no plan files
via an error in the future. For now, create compatibility with v8 to get
it release-ready.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>