🐛 update provider defaults + missing connection types + v8 compat
Connection types were missing and thus caused any request that would
look for them to fail. In most cases this is not noticable at all
(because you eg connect to github to continue to do things with github
as you discover more stuff).

But there is a very nasty case where this hurt: When we are in serve
mode and don't have a primary connection that triggers provider
installations. In this case we receive asset types (or rather v8 Backend
fields!), which then didn't find any way to look up their provider.

So:
1. Add a script to help with defaults generation (it does 80% of the
   work, you only have to copy it over):
   ```bash
   version defaults providers/*/
   ```
2. Update all providers with it
3. Support v8 `Backend` field (needs in-depth testing since both fields
   are not fully compatible; add translations where necessary)

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>