⭐ support ephemeral vs normal runtimes
Ephemeral runtimes are used as throw-away runtimes, that come with their
own plugin runner (and execution space). These runtimes can be thrown
away after they are used, while other runtimes and plugins remain
shared.

This is particularly useful when scanning targets with the discover
mode, where potentially many runtimes/plugins need to be instantiated on
the fly and are only required for the lifetime of the scan.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>