🐛 switch providers after CLI parsing
Early runtime checks can cause providers to switch. For example: A user
may start with the `aws` provider, but really want to user the `os`
provider to scan an instance in AWS:

```bash
cnquery shell aws ec2 instance-connect ec2-user@i-1234abc --region us-east-1 --profile home
```

In the above example, we use the `aws` provider to parse the CLI.
However, after that it's all about handing off the actual asset to the
`os` provider to execute the scan.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>