🧹 make providers/install/X for inidivual providers (#2210)
This has been bugging me, since I often build + install providers as a
santiy check before releases (i.e. not as a builtin; let's call it
overly cautious).

The new structure mirrors `make providers/build/X` and is a nice
complement. Only annoying part is having to line the extra lines for
each provider, but it's also easy and straightforward to copy.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>