✨ retain host scheme in network provider (#2221)
The network provider had a small workaround for schemes to guess if we
were dealing with HTTP or HTTPS. However, this is not safe to use
outside of the two standard ports (80=http, 443=https).

I noticed that the `Runtime` field was unused for network connections.
This change uses this field to retain the "runtime of the network
connection". Users can choose to specify or not specify it. The host
provider works perfectly well without specifying it of course.

Examples:

```coffee
> cnquery run host mondoo.com -c "http.get { statusCode version url }"
http.get: {
  statusCode: 200
  url: url string="http://mondoo.com:80"
  version: "2.0"
}
```

With scheme, default ports are guessed:

```coffee
> cnquery run host https://mondoo.com -c "http.get { statusCode version url }"
http.get: {
  statusCode: 200
  url: url string="https://mondoo.com:443"
  version: "2.0"
}
```

And the effect of using a scheme with the wrong port:

```coffee
> cnquery run host https://mondoo.com:80 -c "http.get { statusCode version url }"
1 error occurred:
	* Get "https://mondoo.com:80": http: server gave HTTP response to HTTPS client
```

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>