🐛 provider installation should refresh schema
Extensible schemas are used to go across multiple providers and
aggregate a schema. We use them extensively, in order to load all
available resources and fields. However, if a new provider is installed
after a schema has been fully loaded, it won't show up, because the
schema doesn't know it has to refresh.

This is now fixed. We keep track of the last time a provider was
installed to the system. Whenever that happens, a timestamp is
updated. Each extensible schema keeps track of this timestamp. If it
changes, the schema gets refreshed.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>