🐛 provider installation should refresh schema (#2305)
* 🐛 provider installation should refresh schema

Extensible schemas are used to go across multiple providers and
aggregate a schema. We use them extensively, in order to load all
available resources and fields. However, if a new provider is installed
after a schema has been fully loaded, it won't show up, because the
schema doesn't know it has to refresh.

This is now fixed. We keep track of the last time a provider was
installed to the system. Whenever that happens, a timestamp is
updated. Each extensible schema keeps track of this timestamp. If it
changes, the schema gets refreshed.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

* 🐛 make sure provider installation clear cache+time

This is happening in the wrong position. It's not just installVersion,
but other paths like InstallFile or even calling InstallIO directly.

Moved to InstallIO.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

* 🟢 do not init last install to 0

otherwise we'd have to force the extensible schema to learn an init
method, this is much more effective

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>