✨ Improve default host handling for no scheme/port
Users sometimes target a system like this:

```
cnquery shell host mondoo.com
```

Question is: What is this? If we connect to it, what do we connect to?

Until now, we'd opt to HTTP, which was of course nice for `http.get`.
But it was horrible for `tls`, because it tried to connect on the
default HTTP port - definitely not what the user intended.

This new mechanism allows users to skip the scheme and the port and ask
the resources to figure out what they want to do with it. For `http`
this is something different than for `tls`.

Example:

```
cnquery shell host mondoo.com
```

with the queries:

```coffee
> http.get { version url.string }
http.get: {
  url.string: "http://mondoo.com"
  version: "2.0"
}

> tls { socket versions }
tls: {
  versions: [
    0: "tls1.2"
    1: "tls1.3"
  ]
  socket: socket protocol="tcp" port=443 address="mondoo.com"
}
```

As you can see, both resources change the port, based on what users
intended. Users can still override this by providing a proper URL or
specifying the port.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>