✨🐛 support switching providers post-connect (#2187)
* ✨🐛 support switching providers post-connect

In the original design, the connect call would always return the
connected asset. However, we have some legacy code that behaved
differently in v8. When it was translated to v9, we had to return the
fully discovered asset, which is the asset that we actually want to
connect.

For now, the connect call will continue to fulfill two functions:
1. discover the actual asset and
2. establish the connection

During this time, we will check once if the provider changed
after the connect call and switch providers if that happened.

We want a distinct discover and connect call in the future.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Signed-off-by: Christian Zunker <christian@mondoo.com>
Co-authored-by: Christian Zunker <christian@mondoo.com>