✨ introduce provider heartbeats
This introduces heartbeats to providers as merged in https://github.com/mondoohq/cnquery/pull/2381

New providers can still run on old cnquery/cnspec without the heartbeat.

On new versions of cnquery/cnspec (which will be released shortly) it will be mandatory to have a provider with this new feature enabled. We found it has far-reaching implications on process/memory cleanup and thus enforce the update.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>