🧹 add CompilerConfig on all compile-related calls
Until now, the only thing developers can configure when it comes to
compile-related calls is the schema.

However, there are use-cases where we may want to configure other things
in the compiler, like:
- which feature-flags are enabled/disabled
- use of stats

I ran into the last use-case myself, which would have caused large
copy-paste rewrites of the Bundle-compiler, just in order to get access
to all the stats.

At this point it is a more comprehensive experience to not move the
schema around, but instead wire the compiler config directly through all
these calls. This means that more future-facing features will be
configurable and that a chain of calls (like `Bundle.Compile`) uses a
consistent set of compiler configs.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>