🐛 Keep idDetectors across connects
With local connections, we saw the case in cnspec, that on first connect, assets had platformIDs as expected.
With local connections this should only be the hostname and perhaps cloud data.

But on later connects, the set id detectors were emtpy and another default kicked in:
https://github.com/mondoohq/cnquery/blob/main/providers/os/provider/detector.go\#L39-L41

This added the ssh hostkey detector to the asset which is not expected.

Signed-off-by: Christian Zunker <christian@mondoo.com>