🐛 Fixes type error in aws.acm.certificates
Fixes the query:
```
aws.acm.certificates { * }
```

A lot of code is partially creating resources. Once this happens, it
doesn't look like anything is going to come back and fill in the data
that was not fetched based on the definition:
```
private aws.acm.certificate @defaults("arn domainName") {
  // ARN for the certificate
  arn string
  // Time before which the certificate is not valid
  notBefore time
  // Time after which the certificate is not valid
  notAfter time
  // Time when the cert was requested
  createdAt time
  // FQDN for the certificate
  domainName string
  // Status of the certificate (issued, expired, revoked, etc.)
  status string
  // Name of the entity associated with the public key in the certificate
  subject string
  // Retrieves an Amazon-issued certificate and its certificate chain
  certificate() network.certificate
  // Tags associated with the certificate
  tags map[string]string
}
```

This means the resource needs to have its fields lazy initialized or be
completely initialized.

Fixes #2558