🐛 fix contains on nil dicts (#2720)
Any dict entry can be `null`, which means that subsequent calls to its
fields can lead to unexpected behavior. Such is the case of `contains`.

This call makes a lot of sense on top of e.g. a string array:

```coffee
["a", "b"].contains("b")
```

But if the structure doesn't exist, then the call fails in unexpected
ways:

```
left side of operation is null
```

This approach still uses an old way of addressing the problem, which
looks quite a bit ugly in the implementation. We have to check for
non-nil and for non-zero length. Once v10 comes around, we can rely on
the `empty` keyword and all this is reduced to just the internal
non-empty check.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>