🧹 support sysinfo without providers (#2969)
* 🧹 support sysinfo without providers

Normally we require to download and install the OS provider when we want to gather the OS info. This is now no longer necessary. Instead we grab a few very isolated libraries from the OS provider itself and embed them into the runtime. This yielded no significant increase in the runtime (500kb). While we want to further decrease this size over time, it is important to support scanning other provider targets without requiring the OS provider to be pulled.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

* 🧹 decouple os connection type constant
* 🧹 extract os filesystem connection into submodule

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>