🐛 prevent testutils from loading local providers (#3040)
Testutils are meant as an in-memory testing utility for unit tests.
However, in their default configuration they pull providers from the
local system, which can have unforseen consequences.

Most of the code in `testutils.go` already sets up the runtime to use
in-memory providers with linked dependencies.

This changes ensures that we don't accidentally load installed providers
on the system - or try to install online providers. For unit tests, we
rely on providers that are fast to grab and execute, thus executable
from memory. Other types of testing (particularly smoke/integration) are
a different matter and not in scope of these utilities at the time of
writing. (Which doesn't mean that they can't become that, but I think we
want to be more explicit about when that happens, to avoid accidental
side-effects where unit tests start to install providers)

Counter-points to this are very welcome!

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>