🐛 Fix port resource panic (#3076)
It is possible for the port resource to panic on mac and windows if the
process is requested. This can happen because we query for all the
processes first, then we query for ports which gets us process ids. A
new process showing up between the first and second queries can cause
the panic to happen.

Linux has a similar check