make sure the os provider can be cross referenced from another provider
this is needed by the k8s provider since it is referencing the containerImage resource, which is defined in the os provider

we also have the issue where filters that shouldn't be executed for an asset may request a provider that shouldn't be called. For example, trying to run aws filters on a k8s asset. This results in printing too many error messages, so I changed the log level of that error to debug

Signed-off-by: Ivan Milchev <ivan@mondoo.com>