🧹 ensure we grab the non-boot volume when multiple xfs uuid volumes (#3762)
we've seen this on a few machines including almalinux.
the boot volume is not labeled as a boot volume. we've consistently seen in this
case that the last of the matching entries is the volume we want. this slight
change in logic allows us to ensure we get the last matching entry of the entries.