🧹 default to MQL asset context
We had this feature-flag since v7, aimed to make it the default in v8. Until now, it has not been mainlined.

This is changing today. Compiling with context is generally very helpful
and provides nice auto-complete on embedded resources (similar to how
you can access embedded fields in golang structs).

There is more work to be done, namely that we want to tie the context of
the connection to the MQL execution itself. This will allow us to call
fields without constantly refering to the global context in the future.
(i.e. instead of `docker.file.sth`, once you connect to a Dockerfile you
can just call `sth`). For this reason, we retain the feature-flag and
will complete its functionality in this major release.