🧹 default to MQL asset context (#3805)
* 🧹 default to MQL asset context

We had this feature-flag since v7, aimed to make it the default in v8. Until now, it has not been mainlined.

This is changing today. Compiling with context is generally very helpful
and provides nice auto-complete on embedded resources (similar to how
you can access embedded fields in golang structs).

There is more work to be done, namely that we want to tie the context of
the connection to the MQL execution itself. This will allow us to call
fields without constantly refering to the global context in the future.
(i.e. instead of `docker.file.sth`, once you connect to a Dockerfile you
can just call `sth`). For this reason, we retain the feature-flag and
will complete its functionality in this major release.

* 🟢 remove cross-provider aliases (for now)

These were added in v7, but unfortunately we had to break the compatibility in v9. We will regain this ability, once cross-provider calls (in the new model) are added back.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>