🐛 fix int-map key types
We currently don't have deep support for int-maps, since we don't have
any use-cases for them yet. That said, they have been supported in the
underlying stack for when we need them. Unfortunately a few of the
supporting elements have been out of sync.

This change aligns data conversions in llx to use int64 maps. Ditto for
`isTruthy` in rawdata. There are a few more to tackle...

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>