Make the yaml file for docs generation required
The way the filestat was being checked if a user gave an incorrect path
we'd love that the file was not provided and move on. These should
always be provided so let's just not include this check entirely.

I also improved a bit of text while I was in here and fixed incorrect
help

Signed-off-by: Tim Smith <tsmith84@gmail.com>