🐛 prevent hang when HelloRetryRequest is returned (#4036)
It was possible for the following query to hang
```
tls.ciphers
```

This happened in cases where the server response with a
`HelloRetryRequest`. This looks like a `ServerHello` until you check the
random field of that message.

We ended up hanging because we thought the server replied successfully
with a hello and were waiting for more data, and the server was waiting
for us to write.