🧹 Do not expire the cache for github provider
We do not have a long running server where we need to clear our cache
every once in a while, instead we execute a single scan and we expect it
to finish (end execution) which ultimately clears the cache. For those
long running executions, we want to persist the cache until we finish
scanning.

This change sets the github provider cache to No expiration.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>