⚡️ increase github expiration cache to 24h (#4799)
Long running scan executions ask for the same resources multiple times
since we expire the github provider cache every 30 minutes. This change
is increasing the time to 24 hours. This will avoid a single scan to ask
for the same resources multiple times, this way we will decrease the
execution time of long running scans.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>