🛠️ new util/zerologadapter for retryable requests
We use the `retryablehttp` package to retry http requests and we always
want the logs to go to `debug` level so, we need an adapter.

We have implemented this adapter many times, this change removes all the
duplicated code and creates a util package that all providers can use.

Example:

```go
retryClient := retryablehttp.NewClient()
retryClient.Logger = zerologadapter.New(log.Logger)
```

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>