⚡️ verify github connection only once
When a user scans a single GH repository, this was not a problem, but
for users that scan an entire organization with thousands of
repositories, we were creating a connection for each asset and
validating it thousands of times.

This changes the behavior to verify the connection only once which will
save us thousands of GH API requests.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>