Unify more of the aws resources on createdAt
createdAt is what Amazon uses in most of their SDK. Some of the older
APIs use different values and in many cases we copied those over into
our resources. This continues the work to unify all of the AWS resources
on createdAt so it's easier to explore fields on resources without
looking at the docs.

Signed-off-by: Tim Smith <tsmith84@gmail.com>