✨ backup way to describe aws regions
This is a workaround for cases where the DescribeRegions API is blocked.
We use a well known regional table provided by AWS to list all possible
regions.

https://api.regional-table.region-services.aws.a2z.com/index.json

NOTE: if we need to validate that we have access to that region or, that
the region is enabled, we can improve this function to do STS identity
calls for all regions.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>